requirements/base.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-filter==24.1
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
